python/requirements.txt
requests
openpyxl
python-dotenv